msm: camera: sensor: Optimize the sensor lite remote flush cmd packet
-This change optimizes the sensor lite remote flush cmd packet and
 Fixes the error response of remote flush command

CRs-Fixed: 3577569

+++ drivers/cam_sensor_module/cam_sensor_utils/cam_sensor_lite_ext_headers.h
@@ -683,8 +683,8 @@ struct host_dest_camera_init_payload_v3 {
  */
 struct sensorlite_sys_cmd {
 	struct sensor_lite_header            header;
-	uint32_t                             sensor_id;
-	uint16_t                             sys_cmd;
+	uint8_t                              sensor_id;
+	uint8_t                              sys_cmd;
 	uint16_t                             reserved;
 } __packed;
